refactor(building-components): lift warm LOD host helpers out of partitions
Move structural `LodSceneHost` scaffolding to crate `lod_host`; keep partition mesh-set → host wiring in `partitions/host`. Document the split (and planned fourth ultra-low GLB) on both sides; note that `PartitionNode` covers direct kits and tessellated concepts.

## Agent Dialogue

- **Developer:** Top half is three GLB paths ↔ one host until another domain shares that layout — technically four soon (ultra-low not authored yet). Give the simplification a shot; split module comments on top-level helpers and `partitions/host`; note partition nodes cover direct mappings and tessellated concepts.
- **Agent:** Added `lod_host` (warm mesh/content hosts); slimmed `partitions/host` to resolution policy; comments for four-tier / ultra-low sharing Low; `PartitionNode` / partitions module notes for direct + tessellated; tests green.
